requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
grpcio==1.76.0
grpcio-tools==1.76.0
idna==3.11
protobuf==6.33.2
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.2
